Add "--no-self-contained" to dotnet build and dotnet publish commands
Add "--no-self-contained" to dotnet build and dotnet publish commands.
This parameter tells the command to not compile and include the dotnet
runtime together with the compiled source code. The runtime does not
need to be included as it is already present in our dotnet Docker images
(torizon/dotnet, torizon/dotnet-debug, torizon/aspdotnet, etc). For more
information, check the dotnet publish command documentaion:
https://learn.microsoft.com/en-us/dotnet/core/tools/dotnet-publish

Related-to: TIE-716
Signed-off-by: Andre Riesco <andre.riesco@toradex.com>