cppdbg: Add engineLogging info on cppdbg of launch.json
Sometimes an error on the execution of the application may not be
shown on the Debug Console, so it may be interesting to get the full
log by setting this "engineLogging" variable to true. However, it prints
a lot of stuff on Debug Console, so setting this to false as default.

Signed-off-by: Andre Riesco <andre.riesco@toradex.com>