cConsole:Deprecating it for Torizon OS 7(move to C/C++ Console template)
Deprecating it for Torizon OS 7, since the maintainer will no longer be
able to maintain it. #273
Also, C Console template is not needed, as the C/C++ Console template
can also be used with C (information explaining how to use GCC instead
of G++ was added to to the `.doc/README.md` of the C/C++ Console
template).

Signed-off-by: Andre Riesco <andre.riesco@toradex.com>